Fixing an int overflow bug
If a client did more than 2^31 requests to a single cache, they would
overflow and get an index out of bounds error during client selection.

This uses the long counter we already had and takes the absolute value
(just in case someone manages to do 9 quintillion requests) and then
mods before truncating to ensure we use the long value and not a
falsely truncated int value.