Add a metric to track "Wrong key returned" errors
This change adds "wrongKeyReturned" as a metric and increments it
everytime we see it in action.

This change was added as we noticed some of these error messages in
the logs, and we'd like to track the frequence of these errors.

Note that even prior to this change, we did not return the wrong key
back to the calling application but instead count it as a miss.

Print stack trace when we notice a Wrong Key error

Consolidate duplicate code under checkWrongKeyReturned()

Address review comments for wrong key returned metric

- Better logging to catch if returned key is from a different host
- Make sure stack trace goes to logs instead of stdout
- Naming nits addressed