refactor: Extract interfaces for NodeLocator algos
A HashRingAlgorithm defines the algorithm that will be used to split a
node into a ring. It assumes the ring will be Ketama-styled in that one
node can be responsible for multiple keyspace ranges. Hash-rings can
also be 'Simple' when they have a 1:1 mapping (one node receives one
keyspace).

A NodeLocatorLookup carries the data layout and associated search
functions to locate a node for a given hash in the hashring.

We also wire in the EVCacheNodeLocator to use the existing default
implementations.