perf: allow concurrent decompress away from network loop
We want to ensure the transcode passes everything that is expected to be
run asynchronously over to the decode loop. In general, memcached calls
us back with gotData, then receivedStatus, then complete. We use gotData
to launch the work onto the transcode threadpool and return control to
memcached, which would immediately then call receivedStatus.

Previously, receivedStatus and complete were set up to interact and set
a value on the underlying future but only by synchronously blocking for
the transcode future. Given this callback is happening nearly immediately
after the gotData callback, we were firing the transcode and nearly
always performing a blocking get, which triggers a synchronous
decompression on the network IO loop. This is of course very detrimental
to evcache driver performance, since the driver cannot even accept new
requests to issue to them to memcached backends, and must wait until
decompression completes.

In this fix, we rearrange things a little to ensure that if the async
decode is requested, that we push the completion status updates to
happen only after the async decode completes. This is a little ugly
because of the current arrangement of the memcached decoder. A future
change might be to overhaul this integration and pull it out of the
memcached transcode framework and use something a bit more friendly.